refactor(workflow): use long IDs and self-contained SQL Server migration

See merge request barzan/ettadbackend!105

+++ Ettad.RequestManagement.Service/OrderItemTracking/OrderItemTrackingService.cs
@@ -100,7 +100,7 @@ namespace Ettad.RequestManagement.Service.OrderItemTracking
             }
         }
 
-        public async Task RecordFinalApprovalHistoryAsync(long orderId, int workflowApprovalStepId)
+        public async Task RecordFinalApprovalHistoryAsync(long orderId, long workflowApprovalStepId)
         {
             try
             {
