+++ Project.Data/Entities/Workflows/WorkflowStepTransition.cs
@@ -2,10 +2,10 @@ using Ettad.CrossCutting.Comman.Base;
 
 namespace Ettad.Data.Entities.Workflows
 {
-    public class WorkflowStepTransition : AuditEntity<int>
+    public class WorkflowStepTransition : AuditEntity<long>
     {
-        public int SourceWorkflowStepId { get; set; }
-        public int TargetWorkflowStepId { get; set; }
+        public long SourceWorkflowStepId { get; set; }
+        public long TargetWorkflowStepId { get; set; }
 
         public WorkflowStep SourceWorkflowStep { get; set; }
         public WorkflowStep TargetWorkflowStep { get; set; }
